Refactor vacation clustering collaborators
